+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ### Chore & Maintenance
 
-- `[*]` Replace `pretty-format/ConvertAnsi` with `jest-serializer-ansi-escapes` ([#12935](https://github.com/facebook/jest/pull/12935))
+- `[*]` Replace internal usage of `pretty-format/ConvertAnsi` with `jest-serializer-ansi-escapes` ([#12935](https://github.com/facebook/jest/pull/12935))
 
 ### Performance